refactor for authz reconciler
